+++ src/api/hooks/useSubmitTransaction.ts
@@ -1,7 +1,10 @@
 import {TxnBuilderTypes} from "aptos";
 import {useState} from "react";
-import {walleExplorertNetworkMap} from "../../constants";
-import {useWalletContext} from "../../context/wallet/context";
+
+import {
+  useWalletContext,
+  walletExplorertNetworkMap,
+} from "../../context/wallet/context";
 import {useGlobalState} from "../../GlobalState";
 import {signAndSubmitTransaction} from "../wallet";
 
@@ -29,7 +32,7 @@ const useSubmitTransaction = () => {
     payload: TxnBuilderTypes.TransactionPayloadScriptFunction,
   ) {
     // if dApp network !== wallet network => return error
-    if (walleExplorertNetworkMap(walletNetwork) !== state.network_name) {
+    if (walletExplorertNetworkMap(walletNetwork) !== state.network_name) {
       setTransactionResponse({
         succeeded: false,
         message:
