+++ src/context/wallet/context.tsx
@@ -1,5 +1,6 @@
 import {createContext, useContext} from "react";
-import {WalletNetworks} from "../../constants";
+import {NetworkName} from "../../constants";
+import { assertNever } from "../../utils";
 
 export interface walletContext {
   isInstalled: boolean;
@@ -19,4 +20,21 @@ export const useWalletContext = () => {
     throw new Error("useWalletContext must be used within a walletContext");
   }
   return context;
+};
+
+export type WalletNetworks = "Devnet" | "Localhost" | "Testnet";
+
+export const walletExplorertNetworkMap = (
+  walletNetwork: WalletNetworks,
+): NetworkName => {
+  switch (walletNetwork) {
+    case "Devnet":
+      return "devnet";
+    case "Localhost":
+      return "local";
+    case "Testnet":
+      return "test";
+    default:
+      return assertNever(walletNetwork);
+  }
 };
