+++ src/pages/layout/NetworkSelect.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect} from "react";
 import {FormControl, Select, SelectChangeEvent} from "@mui/material";
-import {ExplorerNetworks, networks} from "../../constants";
+import {NetworkName, networks} from "../../constants";
 import {useGlobalState} from "../../GlobalState";
 import {useTheme} from "@mui/material/styles";
 import MenuItem from "@mui/material/MenuItem";
@@ -13,7 +13,7 @@ export default function NetworkSelect() {
   const [state, dispatch] = useGlobalState();
   const [searchParams, setSearchParams] = useSearchParams();
 
-  function maybeSetNetwork(network_name: ExplorerNetworks | null) {
+  function maybeSetNetwork(network_name: NetworkName | null) {
     if (!network_name || state.network_name === network_name) return;
     const network_value = networks[network_name];
     if (network_value) {
@@ -25,14 +25,14 @@ export default function NetworkSelect() {
   const handleChange = (event: SelectChangeEvent) => {
     const selectedNetwork = event.target.value as string;
     if (!selectedNetwork) return;
-    const network_name = selectedNetwork.toLowerCase() as ExplorerNetworks;
+    const network_name = selectedNetwork.toLowerCase() as NetworkName;
     maybeSetNetwork(network_name);
   };
 
   useEffect(() => {
     const selectedNetwork = searchParams.get("network");
     if (!selectedNetwork) return;
-    const network_name = selectedNetwork.toLowerCase() as ExplorerNetworks;
+    const network_name = selectedNetwork.toLowerCase() as NetworkName;
     maybeSetNetwork(network_name);
   });
